Elzwelle_2_Stopwatch/requirements.txt
fastapi
uvicorn
jinja2
markupsafe==2.0.1
ttkbootstrap
aiomqtt
pyserial